blog_api/requirements.txt
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.1
pycodestyle==2.5.0
pylint==2.5.2
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
